requirements.txt
fastapi
uvicorn
opensearch-py
python-docx
pymupdf
openpyxl
xlrd
python-pptx
python-dotenv
